Created Docs and Help
Created a Documentation and Help file in each Package. These are intended as a compilation of WIP on things in the relevant classes, notes on easily-forgotten quirks of relevant classes, and a list of the functions of each class in the package. Feel free to expand and edit these documents.